feat: rename `blip` command to `caption` ensure generic
